+++ iPhone/Classes/Government/GovExpenseTypesViewController.m
@@ -167,13 +167,13 @@
 
 -(BOOL) enoughTextChange:(NSString*) sText
 {
-	int sLen = [sText length];
-	int cLen = self.searchText == nil? 0 : [self.searchText length];
+	NSUInteger sLen = [sText length];
+	NSUInteger cLen = self.searchText == nil? 0 : [self.searchText length];
 	if (sLen > cLen + 2 || sLen < cLen -2)
 		return YES;
 	
-	int minLen = sLen > cLen? cLen:sLen;
-	int maxLen = sLen > cLen? sLen:cLen;
+	NSUInteger minLen = sLen > cLen? cLen:sLen;
+	NSUInteger maxLen = sLen > cLen? sLen:cLen;
 	int firstDiffPos = 0;
 	for (int firstDiffPos = 0; firstDiffPos <minLen; firstDiffPos++)
 	{
